ddpg: functional implementation
Getting about 3-3.5k reward on HalfCheetahV3, which is lower than reported for
this flavor of DDPG in the TD3 paper, so either there is some difference between
HalfCheetahV1 and HalfCheetahV3 or this implementation is off somehow. To sanity
check this, I will implement TD3 and see if there is a discrepancy between
rewards and the reported TD3 rewards in the paper.